refactor: extract frontend UI components
