+++ work/.gitignore
@@ -3,3 +3,4 @@ dist/
 *.tsbuildinfo
 *.log
 .DS_Store
+CONVERSATION.private.md
