+++ work/CONVERSATION.md
@@ -1,0 +1,148 @@
+# 项目对话记录
+
+> Codex 对话 ID：`019ed52f-eca4-70a1-9381-9a2e3a6f3d2e`  
+> 整理日期：2026-06-27  
+> 安全说明：服务器密码、GitHub Token 等敏感凭据均已脱敏，不会写入项目。
+
+## 1. 作业背景
+
+**用户：**
+
+这是我的面试作业，我是一名前端，请查看作业内容并给出计划。
+
+作业仓库：
+
+`feedmepos/se-take-home-assignment`
+
+**Codex：**
+
+阅读了仓库 README，确认作业要求实现一个麦当劳订单控制器，前端或后端二选一。
+
+考虑到候选人是前端工程师，建议选择前端实现，技术栈为：
+
+- React
+- TypeScript
+- Vite
+- Vitest
+- React Testing Library
+
+主要功能包括：
+
+- 新增普通订单。
+- 新增 VIP 订单。
+- VIP 订单排在所有普通订单前面。
+- 同类型订单保持先进先出。
+- 增加或减少机器人。
+- 每个机器人同一时间处理一笔订单。
+- 每笔订单需要 10 秒完成。
+- 删除正在工作的最新机器人时，将订单放回等待队列。
+- 展示等待、处理和完成状态。
+
+## 2. 开发计划
+
+**Codex：**
+
+项目按以下步骤实施：
+
+1. 初始化 React + TypeScript + Vite 项目。
+2. 将订单调度规则抽成独立的纯 TypeScript 模块。
+3. 使用 reducer 管理订单、机器人、完成记录和事件日志。
+4. 实现等待区、机器人区和完成区三栏界面。
+5. 编写 VIP 优先级、机器人调度和订单回队列相关测试。
+6. 构建生产版本。
+7. 使用 Docker 和 Nginx 部署到公开服务器。
+
+## 3. 服务器部署
+
+**用户：**
+
+提供了一台带有 Docker 的服务器，并授权直接部署。
+
+**Codex：**
+
+服务器凭据已用于临时 SSH 会话，但未写入任何项目文件。
+
+部署结果：
+
+- Docker 容器名称：`feedme-order-controller`
+- Docker 镜像：`nginx:alpine`
+- 宿主机端口：`13200`
+- 容器端口：`80`
+- 静态文件路径：`/var/www/feedme-order-controller`
+- 线上地址：`http://38.244.31.190:13200/`
+
+部署后通过 HTTP 请求确认返回 `200 OK`。
+
+## 4. 第一版实现结果
+
+完成内容：
+
+- React + TypeScript + Vite 前端应用。
+- 普通订单和 VIP 订单创建。
+- 订单编号唯一递增。
+- VIP 优先级队列。
+- 同类型订单先进先出。
+- 机器人增加、删除和自动取单。
+- 订单处理 10 秒后进入完成区。
+- 删除最新机器人时，将处理中订单放回等待队列。
+- 等待区、机器人区、完成区和事件日志。
+- 核心调度逻辑单元测试。
+
+验证结果：
+
+- 6 个单元测试全部通过。
+- TypeScript 编译通过。
+- Vite 生产构建通过。
+- 本地浏览器交互验证通过。
+- 线上服务返回 `200 OK`。
+
+## 5. 中文化与代码质量调整
+
+**用户：**
+
+要求检查代码质量、修复事件数量样式问题，将 UI 和 README 改成中文，并增加便于讲解的注释。
+
+**Codex：**
+
+完成以下调整：
+
+- 页面标题、按钮、状态、空状态和事件日志改成中文。
+- 核心调度日志改成中文。
+- README 改成中文，并补充状态模型和主要业务规则。
+- 为订单分配、完成判断和队列排序增加中文注释。
+- 修复事件数量徽标宽度不足导致文本溢出的问题。
+- 徽标增加水平内边距和 `white-space: nowrap`。
+- 将倒计时进度条改为根据真实毫秒时间计算，使动画更连续。
+- 修改后再次运行测试和构建，均通过。
+
+## 6. 核心设计说明
+
+订单控制器位于：
+
+`src/domain/orderController.ts`
+
+React 页面主要负责：
+
+- 接收用户操作。
+- 向 reducer 发送 action。
+- 展示订单和机器人状态。
+- 定时触发 `tick` action。
+
+调度模块主要负责：
+
+- 创建订单。
+- 创建和删除机器人。
+- 按 VIP 优先级和创建顺序排序。
+- 给空闲机器人分配订单。
+- 判断订单是否处理满 10 秒。
+- 将完成订单移入完成区。
+- 将被中止的订单重新放回等待队列。
+
+这种拆分让业务规则可以脱离 UI 单独测试，也更适合在面试时讲解。
+
+## 7. 安全提醒
+
+- 不要将服务器密码或 GitHub Token 提交到 Git。
+- 建议服务器使用 SSH Key 登录。
+- 已经在聊天中暴露过的密码和 Token 应尽快轮换。
+- 部署环境中的敏感配置应使用环境变量或服务器密钥管理方式保存。
