+++ work/vite.config.js
@@ -1,0 +1,5 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+export default defineConfig({
+    plugins: [react()],
+});
